refactor: move updateLocationInformation back to file-report.js
Following maintainer feedback:
- Relocated 'updateLocationInformation' from 'lib/shared/location-utils.js' back to 'lib/linter/file-report.js'.
- Exported the function from 'file-report.js'.
- Updated 'lib/linter/linter.js' to import it from './file-report'.
- Removed the now-redundant 'lib/shared/location-utils.js'.
- Updated tests to reflect the new import path.
