+++ lib/linter/linter.js
@@ -46,8 +46,7 @@ const { containsDifferentProperty } = require("../shared/option-utils");
 const { Config } = require("../config/config");
 const { WarningService } = require("../services/warning-service");
 const { SourceCodeTraverser } = require("./source-code-traverser");
-const { FileReport } = require("./file-report");
-const { updateLocationInformation } = require("../shared/location-utils");
+const { FileReport, updateLocationInformation } = require("./file-report");
 
 //------------------------------------------------------------------------------
 // Typedefs
